use float for all quantities

+++ src/conf.go
@@ -9,15 +9,15 @@ import (
 )
 
 type Configuration struct {
-	LogPath string `yaml:"log_path"`
-	DbgMode bool   `yaml:"dbg_mode"`
-	Markup  int    `yaml:"markup"`
-	Port    string `yaml:"port"`
-	WebRoot string `yaml:"web_root"`
-	DBFile  string `yaml:"db_file"`
-	OrgName string `yaml:"org_name"`
-	TLSCert string `yaml:"tls_cert"`
-	TLSPKey string `yaml:"tls_pkey"`
+	LogPath string  `yaml:"log_path"`
+	DbgMode bool    `yaml:"dbg_mode"`
+	Markup  float64 `yaml:"markup"`
+	Port    string  `yaml:"port"`
+	WebRoot string  `yaml:"web_root"`
+	DBFile  string  `yaml:"db_file"`
+	OrgName string  `yaml:"org_name"`
+	TLSCert string  `yaml:"tls_cert"`
+	TLSPKey string  `yaml:"tls_pkey"`
 	cfgFile string
 	cfgPath string
 }
